refactor: 자주 사용할 수 있는 메소드 파일 분리
